Add game room support
Adds a new splash screen that starts with create or join. Create leads to the
role selection menu with the image replaced with a code. Another player can
join this room by clicking join and entering the code, which then takes them to
the role selection screen. Players can also go directly to a room by typing the
code into the url after the site name.